+++ crossterm_winapi/src/structs/mod.rs
@@ -1,7 +1,12 @@
 mod coord;
+mod input;
 mod size;
 mod window_coords;
 
 pub use self::coord::Coord;
+pub use self::input::{
+    ButtonState, ControlKeyState, EventFlags, InputEventType, InputRecord, KeyEventRecord,
+    MouseEvent,
+};
 pub use self::size::Size;
 pub use self::window_coords::WindowPositions;